+++ packages/expo-file-system/ios/Next/FileSystemFile.swift
@@ -75,12 +75,21 @@ internal final class FileSystemFile: FileSystemPath {
 
   // TODO: typedarray, blobs, others support
   func write(_ content: TypedArray) throws {
+    try validateType()
+    try validatePermission(.write)
+    try Data(bytes: content.rawPointer, count: content.byteLength).write(to: url)
   }
 
   func text() throws -> String {
     try validateType()
     try validatePermission(.read)
     return try String(contentsOf: url)
+  }
+
+  func bytes() throws -> Data {
+    try validateType()
+    try validatePermission(.read)
+    return try Data(contentsOf: url)
   }
 
   func base64() throws -> String {